Fixed sample function bug, added markov chain sim
The sample function had a bug that caused it to behave improperly when doing weighted sampling where some of the items had zero weights. I believe I have resolved this bug.

Also added a function, absorb_times, that simulates random walks through a discrete-time Markov chain with one or more absorbing states and returns the states converged to from different starting states on separate trials and the number of steps required to converge. See the function's documentation for a more coherent explanation.
THIS FUNCTION WILL LIKELY GO IN AN INFINITE LOOP IF YOU HAVE AN UNREACHABLE ABSORBING STATE.

Finally, added reachable, which tests if one state of a Markov chain is reachable from another. THIS FUNCTION IS NOT COMPLETE! It cannot determine if a state is truly unreachable, because I haven't added a depth-first search type search functionality to determine if one state can be reached from another by multiple steps. This function is used in the absorb_times function.