linear algebra module- get characteristic polynomial!
Added a module with some fun linear algebra functions. See the documentation. No real practical value here, but many of my functions "show their work" (print out intermediate steps) if you ask them to, which is a nice way to visualize the processes involved.. The most exciting thing here is characteristic_polynom, which uses the horrible recursive determinant-finding algorithm and my polynom class (from prob_calc) to find the characteristic polynomial of a matrix, which can in turn be used to get the eigenvalues and eigenvectors of a matrix.
Other than that, main functions are gaussElim and gramschmidt, which do the things their names imply.